Moved some methods to helper
